refactor(stroke-order): split WriteCharacter.tsx into focused modules (Related to #1935)
Extract inline sub-components and stroke validation from WriteCharacter.tsx (687L):

- WriteCharacterUI.tsx: ConfettiParticles, ProgressDots, StepGuidance, Toast, ZhuyinDisplay (stub)
- StrokeValidator.ts: validateStroke() wrapping isStrokeCorrect — single import point for stroke validation logic
- WriteCharacter.tsx: orchestrator-only shell, no inline sub-components

Canvas hook (useStrokeCanvasRenderer) and state machine (useWriteCharacterMachine) were
already extracted in #1858. This PR completes the split per issue spec.

TDD: 9/9 characterization tests pass before and after refactor.
Tautology guard: breaking error string caused test failure (confirmed).
No new TS errors introduced (4 pre-existing in useStrokeCanvasRenderer + test file).
